requirements.txt
# python3.4+
pyglet==1.3.1
pyshaders==1.4.1
pyGLM